Moved calibration into a folder to only have high level classes in top level.
